+++ SAGESharp/SAGESharp.csproj
@@ -49,11 +49,9 @@
   <ItemGroup>
     <Compile Include="BCLFile.cs" />
     <Compile Include="BHDFile.cs" />
-    <Compile Include="Extensions\ASCIIExtensions.cs" />
-    <Compile Include="Extensions\StreamExtensions.cs" />
-    <Compile Include="Extensions\UIntExtensions.cs" />
     <Compile Include="INIConfig.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="SLB\ASCIIExtensions.cs" />
     <Compile Include="SLB\Internal\EqualityUtils.cs" />
     <Compile Include="SLB\ISLBBinaryReader.cs" />
     <Compile Include="SLB\Identifier.cs" />
@@ -67,6 +65,8 @@
     <Compile Include="SLB\Level\Conversation\Internal\FrameBinaryReader.cs" />
     <Compile Include="SLB\Level\Conversation\Internal\InfoBinaryReader.cs" />
     <Compile Include="SLB\SLBBinaryReaderFactory.cs" />
+    <Compile Include="SLB\StreamExtensions.cs" />
+    <Compile Include="SLB\UIntExtensions.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
