+++ SAGESharp/SLB/Identifier.cs
@@ -1,5 +1,4 @@
-﻿using SAGESharp.Extensions;
-using System;
+﻿using System;
 
 namespace SAGESharp.SLB
 {
