+++ SAGESharp/SLB/Internal/IdentifierBinaryReader.cs
@@ -1,5 +1,4 @@
-﻿using SAGESharp.Extensions;
-using System;
+﻿using System;
 using System.IO;
 
 namespace SAGESharp.SLB.Internal
