+++ SAGESharp/SLB/Level/Conversation/Internal/ConversationBinaryReader.cs
@@ -1,5 +1,4 @@
-﻿using SAGESharp.Extensions;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.IO;
