+++ SAGESharp/SLB/Level/Conversation/Internal/FrameBinaryReader.cs
@@ -1,5 +1,4 @@
-﻿using SAGESharp.Extensions;
-using System;
+﻿using System;
 using System.IO;
 using System.Text;
